break up pg operations in order to increase maintainability
